requirements.txt
asgiref==3.8.1
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.4
docutils==0.21.2
faiss==1.5.3
filelock==3.18.0
google_api_python_client==2.172.0
h2==4.2.0
HTMLParser==0.0.2
importlib_metadata==8.7.0
ipython==9.3.0
ipywidgets==8.1.7
jnius==1.1.0
js==1.0
keyring==25.6.0
langchain==0.3.25
langchain_community==0.3.25
langchain_core==0.3.65
mock==5.2.0
numpy==2.3.0
pandas==2.3.0
Pillow==11.2.1
protobuf==6.31.1
pyodide==0.0.2
pyOpenSSL==25.1.0
python-dotenv==1.1.0
redis==6.2.0
sentence_transformers==4.1.0
simplejson==3.20.1
Sphinx==8.2.3
SQLAlchemy==2.0.41
thread==2.0.5
urllib3_secure_extra==0.1.0
watchdog==6.0.0
xmlrpclib==1.0.1
zstandard==0.23.0
